+++ src/Storage.php
@@ -23,7 +23,13 @@ class Storage {
      * @return string
      */
     public static function format_query_key( $oopi_id ) {
-        return Settings::get( 'id_prefix' ) . $oopi_id;
+        $id_prefix = Settings::get( 'id_prefix' );
+
+        if ( strpos( $oopi_id, $id_prefix ) !== 0 ) {
+            return Settings::get( 'id_prefix' ) . $oopi_id;
+        }
+
+        return $oopi_id;
     }
 
     /**
@@ -41,7 +47,7 @@ class Storage {
     /**
      * Query the WP post id by the given Oopi id.
      *
-     * @param  int|string $id The api id to be matched with postmeta.
+     * @param  int|string $id The Oopi id to be matched with postmeta.
      * @return int|boolean The found post id or 'false' for empty results.
      */
     public static function get_post_id_by_oopi_id( $id ) {
@@ -52,6 +58,31 @@ class Storage {
         $prepared = $wpdb->prepare(
             "SELECT DISTINCT post_id FROM $wpdb->postmeta WHERE meta_key = %s",
             $post_meta_key
+        );
+        // Fetch results from the postmeta table.
+        $results = $wpdb->get_col( $prepared ); // phpcs:ignore
+
+        if ( ! empty( $results ) ) {
+            return (int) $results[0];
+        }
+
+        return false;
+    }
+
+    /**
+     * Query the WP term id by the given Oopi id.
+     *
+     * @param  int|string $id The Oopi id to be matched with termmeta.
+     * @return int|boolean The found term id or 'false' for empty results.
+     */
+    public static function get_term_id_by_oopi_id( $id ) {
+        global $wpdb;
+        // Concatenate the meta key.
+        $term_meta_key = static::format_query_key( $id );
+        // Prepare the sql.
+        $prepared = $wpdb->prepare(
+            "SELECT DISTINCT term_id FROM $wpdb->termmeta WHERE meta_key = %s",
+            $term_meta_key
         );
         // Fetch results from the postmeta table.
         $results = $wpdb->get_col( $prepared ); // phpcs:ignore
@@ -145,33 +176,67 @@ class Storage {
     }
 
     /**
+     * Fetch a WP term by its slug.
+     *
+     * @param string $slug     The term slug.
+     * @param string $taxonomy The taxonomy slug.
+     * @return \WP_Term|null A term object or null if none found.
+     */
+    public static function get_term_by_slug( string $slug, string $taxonomy ) : ?\WP_Term {
+        $terms = get_terms(
+            [
+                'slug'            => $slug,
+                'get'             => 'all',
+                'number'          => 1,
+                'taxonomy'        => $taxonomy,
+                'suppress_filter' => true, // No filtering allowed to get raw objects.
+                'lang'            => '', // For Polylang, ignores language filters.
+            ]
+        );
+
+        if ( is_wp_error( $terms ) || empty( $terms ) ) {
+            return null;
+        }
+
+        return $terms[0];
+    }
+
+    /**
      * Create a new term.
      *
-     * @param  array $term Term data.
-     * @param  Post  $post The current post instance.
+     * @param  Term $term Term data.
+     * @param  Post $post The current Oopi post instance.
      *
-     * @return object|\WP_Error An array containing the `term_id` and `term_taxonomy_id`,
+     * @return array|\WP_Error An array containing the `term_id` and `term_taxonomy_id`,
      *                        WP_Error otherwise.
      */
-    public static function create_new_term( array $term, Post $post ) {
-        $taxonomy = $term['taxonomy'];
-        $name     = $term['name'];
-        $slug     = $term['slug'];
-        // There might be a parent set.
-        $parent = isset( $term['parent'] ) ? get_term_by( 'slug', $term['parent'], $taxonomy ) : false;
+    public static function create_new_term( Term $term, Post $post ) {
+        // If parent's Oopi id is set, fetch it. Default to 0.
+        $parent    = $term->get_parent();
+        $parent_id = $parent ? static::get_term_id_by_oopi_id( $parent ) : 0;
+
         // Insert the new term.
-        $result = wp_insert_term( $name, $taxonomy, [
-            'slug'        => $slug,
-            'description' => isset( $term['description'] ) ? $term['description'] : '',
-            'parent'      => $parent ? $parent->term_id : 0,
-        ] );
+        $result = wp_insert_term(
+            $term->get_name(),
+            $term->get_taxonomy(),
+            [
+                'slug'        => $term->get_slug(),
+                'description' => $term->get_description(),
+                'parent'      => $parent_id,
+            ]
+        );
         // Something went wrong.
         if ( is_wp_error( $result ) ) {
             $err = __( 'An error occurred creating the taxonomy term.', 'oopi' );
-            $post->set_error( 'taxonomy', $name, $err );
+            $post->set_error( 'taxonomy', $term->get_name(), $err );
             return $result;
         }
 
-        return (object) $result;
+        // Identify the Oopi term.
+        $wp_term = get_term( $result['term_id'] );
+        $term->set_term( $wp_term );
+        $term->identify();
+
+        return $result;
     }
 }